refactor(package): add frame package
- move Frame, FinalFrame, NormalFrame to frame package
